requirements.txt
typer==0.4.0
pyyaml==5.4.1
requests==2.26.0
